+++ lib/l10n/app_en.arb
@@ -62,24 +62,8 @@
       }
     }
   },
-  "enterScoreDialogTitle": "Enter Score",
-  "@enterScoreDialogTitle": {
-    "description": "Title text for the enter score dialog."
-  },
-  "enterScoreTeamDropdownLabel": "Team",
-  "@enterScoreTeamDropdownLabel": {
-    "description": "Dropdown label for the enter score dialog to select the team that scored for the end."
-  },
-  "enterScoreScoreDropdownLabel": "Score",
-  "@enterScoreScoreDropdownLabel": {
-    "description": "Dropdown label for the enter score dialog to select the score for the end."
-  },
-  "enterScoreDialogSaveButtonLabel": "Enter",
-  "@enterScoreDialogSaveButtonLabel": {
-    "description": "Button label for the enter score dialog to enter the score for the end."
-  },
-  "editScoreDialogTitle": "Edit Score - End {endNumberName}",
-  "@editScoreDialogTitle": {
+  "scoreInputDialogTitle": "Score - End {endNumberName}",
+  "@scoreInputDialogTitle": {
     "description": "Title text for the edit score dialog.",
     "placeholders": {
       "endNumberName": {
@@ -87,6 +71,10 @@
         "example": "1"
       }
     }
+  },
+  "scoreInputEnterButtonLabel": "Enter",
+  "@scoreInputEnterButtonLabel": {
+    "description": "Button text for score input dialog to submit the score."
   },
   "editScoreTeamDropdownLabel": "Team",
   "@editScoreTeamDropdownLabel": {
