+++ lib/models/models.dart
@@ -1,1 +1,2 @@
 export 'curling_end.dart';
+export 'scoreboard_settings.dart';
