+++ lib/models/scoreboard_settings.dart
@@ -1,0 +1,4 @@
+class ScoreboardSettings {
+  ScoreboardSettings({this.numberOfEnds = 8});
+  int numberOfEnds;
+}
