+++ lib/widgets/app_bar/app_bar.dart
@@ -1,2 +1,4 @@
 export 'app_bar_action_button.dart';
 export 'finish_game_dialog.dart';
+export 'score_input_dialog.dart';
+export 'settings_dialog.dart';
